Adopt a new way to go with the loglevels
        Now users can add "DEFAULT_LOG_LOC" to their Makefile to change the default logging location , in a similar way they can change the "ALT_LOG_LOC" to change the alternative  location ( the location lib-le-log falls to if DEFAULT_LOG_LOC fails).

        Now the previous functions of warn/error/info/debug are removed. The parent function has to a pass a "loglevel"
 ======================= \
   label  | loglevel | message     /
 ======================== \
  | Info      | 0        |  Info :           /
  | Warn    | 1        |  Warn :         \
  | Error    | 2        |  Error :        /
  | Debug  | 3        |  Debug :      \
  ======================== /

so now a function looks like `klog(0,"This was a test message");`
and the output `Info : This was a a test message`

We have also added flag `LOG_WRAP_TAGS`. This flag allows the user to wrap the log tags 
so , 
Info = I: 
Warn = W:
Error = E:
Debug = D: 

This allows to user to maximize the output that can be given  (max is 256 chars).

We have also started to use stack based allocations rather than heap based. 

Lots of rudimentary comments have been removed. Every function should have only one comment which briefs the work done by it, rest should be explained  by the code itself. But all the comments have not been removed , some exist ( but changed ) . They exists to explain the file opening sequence. 

The header <unistd.h> has been moved upwards, because it defines the POSIX_VERSION which is used in many other headers